solved fist 8 questions

+++ B.cpp
@@ -2,15 +2,15 @@
 using namespace std ; 
 
 int main(){
-	int n; 
+	long long int n; 
 	cin>>n;
-	int a[n];
+	long long int a[n];
 
 	for(int i=0;i<n; i++){
 		cin>>a[i]; 
 	}
 
-	int sum=0;
+	long long int sum=0;
 
 	for(int i=0;i<n; i++){
 		sum+=a[i];
